requirements.txt
# Core
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.0

# HTTP & Parsing
httpx>=0.26.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0

# Scheduling
apscheduler>=3.10.0

# AI/ML
groq>=0.4.0
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Dashboard
streamlit>=1.30.0
plotly>=5.18.0
pandas>=2.0.0
